Change int to float which fixes the outline shader. Massive thank you to Flozi for solving this issue for me.

+++ ModelMaterials_GL4/templates/cus_gl4.frag.glsl
@@ -78,7 +78,7 @@ uniform samplerCube reflectTex;		//7
 
 //uniform int shadowsQuality;
 int shadowsQuality = 0;
-uniform int materialIndex;
+uniform float materialIndex;
 
 //uniform vec3 rndVec;
